AUtomatically generate data from compat-table
