+++ packages/babel-compat-data/scripts/data/plugin-bugfixes.js
@@ -1,8 +1,21 @@
+/*
+
+Each entry must be in the following format:
+
+type PluginData = {
+  replaces: string,
+  features: [...nameOfCompatTableFeature]
+}
+
+targets is merged with the data from features, and has the precedence.
+
+*/
+
 module.exports = {
   "bugfix/transform-edge-default-parameters": {
     replaces: "transform-parameters",
-    targets: {
-      edge: ["16", "18"],
-    },
+    features: [
+      "destructuring, parameters / shorthand defaults, arrow function",
+    ],
   },
 };
