+++ packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-13/options.json
@@ -5,8 +5,9 @@
       "debug": true,
       "bugfixes": true,
       "targets": {
-        "edge": "16"
-      }
+        "edge": "13"
+      },
+      "exclude": ["transform-destructuring", "transform-arrow-functions"]
     }]
   ]
 }
