+++ packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-13/output.js
@@ -1,5 +1,5 @@
 (_ref) => {
-  let {
+  var {
     x = 2
   } = _ref;
 };
