+++ packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-13/stdout.txt
@@ -1,0 +1,30 @@
+@babel/preset-env: `DEBUG` option
+
+Using targets:
+{
+  "edge": "13"
+}
+
+Using modules transform: auto
+
+Using plugins:
+  transform-function-name { "edge":"13" }
+  transform-for-of { "edge":"13" }
+  transform-dotall-regex { "edge":"13" }
+  transform-parameters { "edge":"13" }
+  transform-block-scoping { "edge":"13" }
+  transform-new-target { "edge":"13" }
+  transform-exponentiation-operator { "edge":"13" }
+  transform-async-to-generator { "edge":"13" }
+  proposal-async-generator-functions { "edge":"13" }
+  proposal-object-rest-spread { "edge":"13" }
+  proposal-unicode-property-regex { "edge":"13" }
+  proposal-json-strings { "edge":"13" }
+  proposal-optional-catch-binding { "edge":"13" }
+  proposal-optional-chaining { "edge":"13" }
+  transform-named-capturing-groups-regex { "edge":"13" }
+  proposal-nullish-coalescing-operator { "edge":"13" }
+  transform-modules-commonjs { "edge":"13" }
+  proposal-dynamic-import { "edge":"13" }
+
+Using polyfills: No polyfills were added, since the `useBuiltIns` option was not set.
