+++ packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-14/options.json
@@ -5,7 +5,7 @@
       "debug": true,
       "bugfixes": true,
       "targets": {
-        "edge": "15"
+        "edge": "14"
       }
     }]
   ]
