+++ packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-14/stdout.txt
@@ -1,0 +1,28 @@
+@babel/preset-env: `DEBUG` option
+
+Using targets:
+{
+  "edge": "14"
+}
+
+Using modules transform: auto
+
+Using plugins:
+  transform-function-name { "edge":"14" }
+  transform-for-of { "edge":"14" }
+  transform-dotall-regex { "edge":"14" }
+  transform-destructuring { "edge":"14" }
+  transform-async-to-generator { "edge":"14" }
+  proposal-async-generator-functions { "edge":"14" }
+  proposal-object-rest-spread { "edge":"14" }
+  proposal-unicode-property-regex { "edge":"14" }
+  proposal-json-strings { "edge":"14" }
+  proposal-optional-catch-binding { "edge":"14" }
+  proposal-optional-chaining { "edge":"14" }
+  transform-named-capturing-groups-regex { "edge":"14" }
+  proposal-nullish-coalescing-operator { "edge":"14" }
+  bugfix/transform-edge-default-parameters { "edge":"14" }
+  transform-modules-commonjs { "edge":"14" }
+  proposal-dynamic-import { "edge":"14" }
+
+Using polyfills: No polyfills were added, since the `useBuiltIns` option was not set.
