--- packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-15/stdout.txt
@@ -1,25 +1,0 @@
-@babel/preset-env: `DEBUG` option
-
-Using targets:
-{
-  "edge": "15"
-}
-
-Using modules transform: auto
-
-Using plugins:
-  transform-function-name { "edge":"15" }
-  transform-dotall-regex { "edge":"15" }
-  transform-parameters { "edge":"15" }
-  proposal-async-generator-functions { "edge":"15" }
-  proposal-object-rest-spread { "edge":"15" }
-  proposal-unicode-property-regex { "edge":"15" }
-  proposal-json-strings { "edge":"15" }
-  proposal-optional-catch-binding { "edge":"15" }
-  proposal-optional-chaining { "edge":"15" }
-  transform-named-capturing-groups-regex { "edge":"15" }
-  proposal-nullish-coalescing-operator { "edge":"15" }
-  transform-modules-commonjs { "edge":"15" }
-  proposal-dynamic-import { "edge":"15" }
-
-Using polyfills: No polyfills were added, since the `useBuiltIns` option was not set.
