--- packages/babel-preset-env/test/fixtures/bugfixes/edge-default-params-edge-16/stdout.txt
@@ -1,25 +1,0 @@
-@babel/preset-env: `DEBUG` option
-
-Using targets:
-{
-  "edge": "16"
-}
-
-Using modules transform: auto
-
-Using plugins:
-  transform-function-name { "edge":"16" }
-  transform-dotall-regex { "edge":"16" }
-  proposal-async-generator-functions { "edge":"16" }
-  proposal-object-rest-spread { "edge":"16" }
-  proposal-unicode-property-regex { "edge":"16" }
-  proposal-json-strings { "edge":"16" }
-  proposal-optional-catch-binding { "edge":"16" }
-  proposal-optional-chaining { "edge":"16" }
-  transform-named-capturing-groups-regex { "edge":"16" }
-  proposal-nullish-coalescing-operator { "edge":"16" }
-  bugfix/transform-edge-default-parameters { "edge":"16" }
-  transform-modules-commonjs { "edge":"16" }
-  proposal-dynamic-import { "edge":"16" }
-
-Using polyfills: No polyfills were added, since the `useBuiltIns` option was not set.
